PB-369: Integrate third party library to work with DOMDocument
 - update wrapper classes
