+++ app/code/Magento/PageBuilder/Model/Dom/Adapter/DocumentFragmentInterface.php
@@ -19,4 +19,25 @@ interface DocumentFragmentInterface
      * @return bool
      */
     public function appendHTML(string $data): bool;
+
+    /**
+     * Returns the first element matching the specified selector.
+     *
+     * @param string $selector
+     * @return ElementInterface
+     */
+    public function querySelector(string $selector): ElementInterface;
+
+    /**
+     * Returns all elements matching the specified selector.
+     *
+     * @param string $selector
+     * @return HtmlCollectionInterface
+     */
+    public function querySelectorAll(string $selector): HtmlCollectionInterface;
+
+    /**
+     * Returns string content of fragment
+     */
+    public function prop_get_innerText(): string;
 }
